Create override version of cover layout
Anything in your project’s`layouts` directory will override the corresponding files in your theme’s `layouts` directory, as long as the file structure and naming are exactly the same. I like to commit this exacrt duplicate before making any changes to it so that the git record will better capture the changes I do make. In case I need to go back, or want to carry those changes over to a new project