Add the contents grid code
The logic is from contents/single.html, lines 24 and 25, but I made sure to also bring in the needed <div>s with their classes. The other thing I had to do is add `class: grid` to the page YAML for cover.md to get it to render correctly, as there's some logic in the underlying contents-list partial that looks for that class.